pyproject.toml
[tool.poetry]
name = "django-default"
version = "0.1.0"
description = ""
authors = ["oguri <oguri.1729@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-debug-toolbar = "^3.8.1"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
Django = "^4.1.7"
django-allauth = "^0.52.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
